pyproject.toml
[project]
name = "educreate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "groq>=0.23.1",
    "huggingface-hub>=0.30.2",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.2.0",
    "openai>=1.76.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "tiktoken>=0.9.0",
    "transformers>=4.51.3",
]
